[同步保持图片]

+++ java/api/src/main/java/com/exadel/frs/core/trainservice/exel/ExelRow.java
@@ -16,7 +16,7 @@ import java.util.Date;
 @AllArgsConstructor
 public class ExelRow
 {
-    private Integer createTimestamp;
+    private Long createTimestamp;
 
     private String  deviceIdAddress;
