+++ java/common/src/main/java/com/exadel/frs/commonservice/projection/CaputreImgProjection.java
@@ -1,4 +1,4 @@
 package com.exadel.frs.commonservice.projection;
 
-public record CaputreImgProjection(long id, int timestamp, String imgUrl, int device_id) {
+public record CaputreImgProjection(long id, long timestamp, String imgUrl, int device_id) {
 }
