+++ java/common/src/main/java/com/exadel/frs/commonservice/projection/SaveFaceImgProjection.java
@@ -5,7 +5,7 @@ import com.exadel.frs.commonservice.entity.SaveFaceImg;
 
 import java.util.UUID;
 
-public record SaveFaceImgProjection(long subId, int deviceId, int timestamp, String imgUrl, String subImgUrl,  UUID embeddingId, String subjectName,  int gender, int minAge, int maxAge, Float similarity, int boxMinX, int boxMinY, int boxMaxX, int boxMaxY)
+public record SaveFaceImgProjection(long subId, int deviceId, long timestamp, String imgUrl, String subImgUrl,  UUID embeddingId, String subjectName,  int gender, int minAge, int maxAge, Float similarity, int boxMinX, int boxMinY, int boxMaxX, int boxMaxY)
 {
